resolve open comments
