+++ .vscode/launch.json
@@ -7,7 +7,7 @@
     {
       "name": "Launch Server via NPM",
       "request": "launch",
-      "runtimeArgs": ["run-script", "dev"],
+      "runtimeArgs": ["run", "dev"],
       "runtimeExecutable": "npm",
       "skipFiles": ["<node_internals>/**"],
       "type": "pwa-node"
