+++ server/utils/index.ts
@@ -1,0 +1,9 @@
+import { PlaylistVideo } from '..';
+
+export const findPlaylistVideoByUrl = (
+  playlist: PlaylistVideo[],
+  url?: string
+) => {
+  if (!url) return;
+  return playlist.find((video) => video.url === url);
+};
